+++ packages/terra-data-grid/src/subcomponents/ColumnHeaderCell.jsx
@@ -262,8 +262,8 @@ const ColumnHeaderCell = (props) => {
       role="columnheader"
       scope="col"
       aria-sort={sortIndicator}
-      onMouseDown={onColumnSelect ? handleMouseDown : null}
-      onKeyDown={(isSelectable || isResizable) ? handleKeyDown : null}
+      onMouseDown={onColumnSelect ? handleMouseDown : undefined}
+      onKeyDown={(isSelectable || isResizable) ? handleKeyDown : undefined}
       style={{ width: `${width}px`, height: headerHeight, left: cellLeftEdge }} // eslint-disable-line react/forbid-dom-props
     >
       <div