+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 * Added
   * Added test for data grid column resizing.
+  * Added example for Flowsheet Data Grid containing "No Result" cells.
+
+* Fixed
+  * Fixed the `terra-tabs` icon-only examples keyboard navigation issue.
+
+* Added
+  * Added example for Flowsheet Data Grid containing "No Result" cells.
 
 * Fixed
   * Fixed broken links in the Worklist Data Grid documentation in `terra-data-grid` component.
@@ -11,8 +18,10 @@
 ## 1.42.0 - (October 20, 2023)
 
 * Added
+  * Added documentation for the new `terra-table` component.
   * Added test to cover enabling zebra striping for the `terra-table` component.
   * Added test for tab focus for scrollable tables for the `terra-table` component.
+  * Added examples and tests to cover the behavior of interactive elements within a table for `terra-table`.
 
 * Updated
   * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.